refactor: TIO-97: extract confirm dialog into own widget
